backend/requirements.txt
django==4.2.13
djangorestframework==3.15.1
django-cors-headers==4.3.1
sentence-transformers==2.7.0
numpy==1.26.4
requests==2.31.0
selenium==4.21.0
webdriver-manager==4.0.1
python-dotenv==1.0.1
